TraceHawkOSINT/requirements.txt
requests
whois
dnspython
ipinfo
fpdf
pyfiglet
termcolor

